% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGDrivE-Auxiliary.R
\name{AnalyzeQuantiles}
\alias{AnalyzeQuantiles}
\title{Summary Statistics for Stochastic MGDrivE}
\usage{
AnalyzeQuantiles(readDirectory, writeDirectory, doMean = TRUE,
  quantiles = NULL, simTime, numPatch, genotypes, remFiles = FALSE)
}
\arguments{
\item{readDirectory}{Directory to find repetition folders in}

\item{writeDirectory}{Directory to write output}

\item{doMean}{Boolean, calculate mean or not. Default is TRUE}

\item{quantiles}{Vector of quantiles to calculate. Default is NULL}
}
\value{
Writes output to files in writeDirectory
}
\description{
This function reads in all repetitions for each patch and calculates either
the mean, quantiles, or both. User chooses the quantiles, up to 4 decimal places,
and enters them as a vector. (order does not matter)  \cr
}
\details{
Given the readDirectory, this function assumes the follow file structure: \cr
 * readDirectory
   * repetition 1
     * patch 1
     * patch 2
     * patch 3
   * repetition 2
     * patch 1
     * patch 2
     * patch 3
   * repetition 3
   * repetition 4
   * ...
}
