\name{fwritetest-package}
\alias{fwritetest-package}
\alias{fwritetest}
\docType{package}
\title{
\packageTitle{fwritetest}
}
\description{
\packageDescription{fwritetest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fwritetest}
\packageIndices{fwritetest}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fwritetest}

Maintainer: \packageMaintainer{fwritetest}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
